Remove opt_capt from goog.events.removeAll.
Use goog.events.listeners_ as source of truth for
goog.events.removeAll when opt_obj is not specified. While
this is supposedly less performant, goog.events.removeAll() is
not meant to be used often and hence it should not matter if
it is performant or not. This should make it easier to
implement delegation to goog.events.Listenable.

R=nicksantos
DELTA=35 (17 added, 13 deleted, 5 changed)


Revision created by MOE tool push_codebase.
MOE_MIGRATION=6045


git-svn-id: http://closure-library.googlecode.com/svn/trunk@2404 0b95b8e8-c90f-11de-9d4f-f947ee5921c8